+++ src/components/generalPageContent/AboutUs.jsx
@@ -1,6 +1,5 @@
 import React from "react";
 import "./AboutUs.style.css";
-import AddressSearch from "../mandatoryQuestions/AddressSearch";
 
 const AboutUs = () => {
   return (
@@ -8,7 +7,6 @@ const AboutUs = () => {
       <div id="aboutUsContent">
         <h1 data-cy="title">About us</h1>
         <h3>Site currently under construction...</h3>
-        <AddressSearch />
       </div>
     </>
   );
