--- src/components/mandatoryQuestions/AddressSearch.jsx
@@ -1,67 +1,0 @@
-import React, { useState } from "react";
-import PlacesAutocomplete, {
-  geocodeByAddress,
-  getLatLng,
-} from "react-places-autocomplete";
-import scriptLoader from "react-async-script-loader";
-
-const AddressSearch = ({ isScriptLoaded, isScriptLoadSucceed }) => {
-  const [address, setAddress] = useState("");
-  const [coordinates, setCoordinates] = useState({
-    lat: null,
-    lng: null,
-  });
-  const handleSelect = async (value) => {
-    const results = await geocodeByAddress(value);
-    const latLng = await getLatLng(results[0]);
-    setAddress(value);
-    setCoordinates(latLng);
-  };
-
-  const searchOptions = {
-    componentRestrictions: { country: ["SE"] },
-  };
-
-  if (isScriptLoaded && isScriptLoadSucceed) {
-    return (
-      <div>
-        <PlacesAutocomplete
-          value={address}
-          onChange={setAddress}
-          onSelect={handleSelect}
-          searchOptions={searchOptions}
-          shouldFetchSuggestions={address.length >= 2}
-        >
-          {({
-            getInputProps,
-            suggestions,
-            getSuggestionItemProps,
-            loading,
-          }) => (
-            <div>
-              <input {...getInputProps({ placeholder: "Type address" })} />
-              <div>
-                {loading ? <div>...loading</div> : null}
-                {suggestions.map((suggestion) => {
-                  const style = {
-                    backgroundColor: suggestion.active ? "#41b6e6" : "#fff",
-                  };
-                  return (
-                    <div {...getSuggestionItemProps(suggestion, { style })}>
-                      {suggestion.description}
-                    </div>
-                  );
-                })}
-              </div>
-            </div>
-          )}
-        </PlacesAutocomplete>
-      </div>
-    );
-  } else {
-    return <div></div>;
-  }
-};
-export default scriptLoader([
-  `https://maps.googleapis.com/maps/api/js?key=${process.env.REACT_APP_GOOGLE_MAP_API}&libraries=places`,
-])(AddressSearch);
